+++ src/Products/product.entity.ts
@@ -1,0 +1,28 @@
+/* eslint-disable prettier/prettier */
+import { User } from 'src/users/user.entity';
+import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity()
+export class Product {
+  @PrimaryGeneratedColumn()
+  id: number;
+  @Column()
+  name: string;
+  @Column()
+  description: string;
+  @Column()
+  price: number;
+  @Column()
+  createUserId: number;
+  @Column()
+  createdAt: Date;
+  @Column({
+    type: 'timestamp',
+    default: () => 'CURRENT_TIMESTAMP',
+    onUpdate: 'CURRENT_TIMESTAMP',
+  })
+  updatedAt: Date;
+
+  @ManyToOne(() => User, (user) => user.products)
+  user: User;
+}
